Fix compilation errors resulting from updating JDK to 1.7.0_75

+++ src/com/jsoft/pst/CRC.java
@@ -59,7 +59,7 @@ class CRC {
 
 		int crc = 0;
 		for(int i=0; i < size; ++i)
-			crc = CrcTableOffset32[(int)(crc ^ byteBuffer.get()) & 0xff] ^ (crc >>> 8); 
+			crc = CrcTableOffset32[(crc ^ byteBuffer.get()) & 0xff] ^ (crc >>> 8); 
 
 		byteBuffer.position(position);
 		return crc;
